+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Improvements
 
+- 🥳 Imagery now loads and displays in the viewer incrementally.
 - ✨ Updated to use `numba` to fuse multiple numpy imagery operations into a single memory pass.
  
 ### Bug Fixes